requirements.txt
langchain
pinecone-client
pypdf
unstructured
python-dotenv
cohere
